move around Login & withCheckLoggedIn component
