+++ frontend/src/App.tsx
@@ -1,16 +1,18 @@
 import axios from "axios"
 import { Route, Routes } from "react-router-dom"
 import { SWRConfig } from "swr"
-import Home from "./pages/Home"
-import { withCheckLoggedIn } from "./pages/Login"
-import Search from "./pages/Search"
-import IssueDetails from "./pages/IssueDetails"
-import Dashboard from "./pages/Dashboard"
+
+import Home from "pages/Home"
+import Search from "pages/Search"
+import IssueDetails from "pages/IssueDetails"
+import Dashboard from "pages/Dashboard"
+import { withCheckLoggedIn } from "utils/login"
 
 export default function App() {
   const DashboardPage = withCheckLoggedIn(Dashboard)
   const SearchPage = withCheckLoggedIn(Search)
   const DetailsPage = withCheckLoggedIn(IssueDetails)
+
   return (
     <SWRConfig
       value={{
