+++ frontend/src/components/common/LoginButton.tsx
@@ -1,12 +1,7 @@
-import { ComponentType } from "react"
-import { Navigate } from "react-router-dom"
-import Cookies from "js-cookie"
-
 const loginUrl =
   "https://comics.auth.us-east-1.amazoncognito.com/oauth2/authorize?client_id=2ca9tq8ue0rp1n8vc5ckfh0c40&response_type=code&scope=email+openid+phone&redirect_uri=http%3A%2F%2Flocalhost%3A9000%2Fapi%2Fuser%2Flogin%2F"
-const TOKEN = "access"
 
-export const LoginButton = () => {
+const LoginButton = () => {
   return (
     <a
       className={
@@ -19,12 +14,4 @@ export const LoginButton = () => {
   )
 }
 
-export function withCheckLoggedIn(Component: ComponentType) {
-  return () => {
-    const cookie = Cookies.get(TOKEN)
-    if (!cookie) {
-      window.location.href = loginUrl
-    }
-    return !cookie ? <Navigate to={"/"} /> : <Component />
-  }
-}
+export default LoginButton
